Change INT to BIGINT in schema.sql

+++ src/main/resources/db/schema.sql
@@ -17,7 +17,7 @@ SET @OLD_SQL_MODE = @@SQL_MODE, SQL_MODE =
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `products`
 (
-    `id`          INT                     NOT NULL AUTO_INCREMENT,
+    `id`          BIGINT                     NOT NULL AUTO_INCREMENT,
     `name`        VARCHAR(255)            NOT NULL COMMENT 'Nombre del producto',
     `price`       DECIMAL(11, 2) UNSIGNED NOT NULL COMMENT 'Precio del producto',
     `img_url`     VARCHAR(255)            NOT NULL COMMENT 'Url de la imagen del producto.',
@@ -94,10 +94,10 @@ CREATE TABLE IF NOT EXISTS `unit_types`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `products_has_shopping_list`
 (
-    `product_id`        INT                     NOT NULL,
-    `shopping_list_id`  INT                     NOT NULL,
-    `unit_type_id`      INT                     NOT NULL,
-    `units_per_product` INT UNSIGNED            NOT NULL DEFAULT 1 COMMENT 'Unidades por producto.',
+    `product_id`        BIGINT                     NOT NULL,
+    `shopping_list_id`  BIGINT                     NOT NULL,
+    `unit_type_id`      BIGINT                     NOT NULL,
+    `units_per_product` BIGINT UNSIGNED            NOT NULL DEFAULT 1 COMMENT 'Unidades por producto.',
     `total_calories`    DECIMAL(11, 2) UNSIGNED NOT NULL COMMENT 'Total de calorías por unidad.',
     `total_price`       DECIMAL(11, 2) UNSIGNED NOT NULL COMMENT 'Precio total por unidad.',
     `selected`          TINYINT(1)              NOT NULL DEFAULT 0 COMMENT 'Establece si el producto esta seleccionado.',
